cats: update subscription views cast to bigint

+++ core/src/cats/ddl/updates/postgresql.2240_2250.sql
@@ -11,42 +11,42 @@ SELECT
   f.fileset AS fileset,
   j.jobbytes::bigint / 1000000 AS total_mb,
   CASE
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%mssqlvdi:%' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-mariabackup%' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-percona%' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-postgres%' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-ldap%' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN NULL::bigint
-  END AS db_mb,
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%mssqlvdi:%' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-mariabackup%' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-percona%' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-postgres%' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-ldap%' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN NULL
+  END::bigint AS db_mb,
   CASE
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-vmware%' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-ovirt%' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN NULL::bigint
-  END AS vm_mb,
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-vmware%' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-ovirt%' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN NULL
+  END::bigint AS vm_mb,
   CASE
-    WHEN f.filesettext ILIKE '%{%{%meta%=' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-libcloud%' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-qumulo%' THEN j.jobbytes::bigint / 1000000
-    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN NULL::bigint
-  END AS filer_mb,
+    WHEN f.filesettext ILIKE '%{%{%meta%=' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-libcloud%' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-qumulo%' THEN j.jobbytes / 1000000
+    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN NULL
+  END::bigint AS filer_mb,
   CASE
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN j.jobbytes::bigint / 1000000
-  END AS normal_mb,
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN j.jobbytes / 1000000
+  END::bigint AS normal_mb,
   CASE
-    WHEN f.filesettext ILIKE '%{%{%meta%=' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%mssqlvdi:%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-mariabackup%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-percona%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-postgres%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-ldap%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-vmware%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-ovirt%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-libcloud%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-qumulo%' THEN NULL::bigint
-    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN NULL::bigint
-    ELSE j.jobbytes::bigint / 1000000
-  END AS unknown_mb,
+    WHEN f.filesettext ILIKE '%{%{%meta%=' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%mssqlvdi:%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-mariabackup%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-percona%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-postgres%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-ldap%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-vmware%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-ovirt%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-libcloud%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%Plugin%=%python%:module_name=bareos-fd-qumulo%' THEN NULL
+    WHEN f.filesettext ILIKE '%{%{%File%=%' THEN NULL
+    ELSE j.jobbytes / 1000000
+  END::bigint AS unknown_mb,
   CASE
     WHEN f.filesettext ILIKE '%{%{%meta%=' THEN NULL
     WHEN f.filesettext ILIKE '%{%{%Plugin%=%mssqlvdi:%' THEN NULL
@@ -91,9 +91,9 @@ SELECT
   NULL::bigint AS vm_units,
   CEIL(SUM(filer_mb) / 1000000)::bigint AS filer_units,
   CASE
-    WHEN SUM(normal_mb) <= 10000000 THEN 1::bigint
-    ELSE CEIL(SUM(normal_mb)/1000000)::bigint - 9::bigint
-  END AS normal_units
+    WHEN SUM(normal_mb) <= 10000000 THEN 1
+    ELSE CEIL(SUM(normal_mb)/1000000) - 9
+  END::bigint AS normal_units
 FROM latest_full_size_categorized
 WHERE normal_mb IS NOT NULL
    OR filer_mb IS NOT NULL
@@ -103,12 +103,12 @@ SELECT
   client,
   fileset,
   CASE
-    WHEN db_mb > 0 THEN 1::bigint
-  END AS db_units,
+    WHEN db_mb > 0 THEN 1
+  END::bigint AS db_units,
   CASE
-    WHEN vm_mb > 10000000 THEN CEIL(vm_mb / 1000000)::bigint - 9::bigint
-    WHEN vm_mb > 0 THEN 1::bigint
-  END AS vm_units,
+    WHEN vm_mb > 10000000 THEN CEIL(vm_mb / 1000000) - 9
+    WHEN vm_mb > 0 THEN 1
+  END::bigint AS vm_units,
   NULL::bigint AS filer_units,
   NULL::bigint AS normal_units
 FROM latest_full_size_categorized
